fix downgrade for redshift

+++ backend/migrations/versions/852cdf6cf1e0_add_redshift_datasets.py
@@ -147,5 +147,6 @@ def downgrade():
         )
         for policy in policies:
             for permission in policy.permissions:
-                session.delete(permission)
-                session.commit()
+                if permission.name in ENVIRONMENT_REDSHIFT_ALL:
+                    session.delete(permission)
+                    session.commit()
